--- HW13/Assets.xcassets/custom.heart.symbolset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "symbols" : [
-    {
-      "filename" : "custom.heart.text.square.svg",
-      "idiom" : "universal"
-    }
-  ]
-}
