--- HW13/Assets.xcassets/custom.magnifyin.symbolset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "symbols" : [
-    {
-      "filename" : "custom.magnifyingglass.svg",
-      "idiom" : "universal"
-    }
-  ]
-}
