--- HW13/Assets.xcassets/custom.square.symbolset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "symbols" : [
-    {
-      "filename" : "custom.square.stack.svg",
-      "idiom" : "universal"
-    }
-  ]
-}
