--- HW13/Assets.xcassets/photo.symbolset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "symbols" : [
-    {
-      "filename" : "custom.photo.fill.on.rectangle.fill.svg",
-      "idiom" : "universal"
-    }
-  ]
-}
