+++ HW13/Resurses/Assets.xcassets/custom.heart.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "custom.heart.text.square.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
