+++ HW13/Resurses/Assets.xcassets/custom.magnifyin.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "custom.magnifyingglass.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
