+++ HW13/Resurses/Assets.xcassets/custom.square.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "custom.square.stack.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
