+++ HW13/Resurses/Assets.xcassets/photo.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "custom.photo.fill.on.rectangle.fill.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
