+++ HW13/Sourses/App/ViewController.swift
@@ -1,0 +1,18 @@
+//
+//  ViewController.swift
+//  HW13
+//
+//  Created by Александр Петрович on 21.12.2021.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+
+
+}
+
