--- HW13/ViewController.swift
@@ -1,18 +1,0 @@
-//
-//  ViewController.swift
-//  HW13
-//
-//  Created by Александр Петрович on 21.12.2021.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-
-
-}
-
