--- HW13/ViewControllers/FirstTabViewController.swift
@@ -1,27 +1,0 @@
-//
-//  FirstTabViewController.swift
-//  HW13
-//
-//  Created by Александр Петрович on 27.12.2021.
-//
-
-import UIKit
-
-class FirstTabViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view.backgroundColor = .systemGray6
-    }
-    
-    override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
-        super.init(nibName: nil, bundle: nil)
-    }
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-    }
-    
-    
-
-}
